refactor(app): 토스트 수명주기를 전용 훅으로 분리
App 안에 남아 있던 토스트 상태, 타이머 정리, exit 전환, exit 완료 처리 로직을 useToastState로 이동했다.

이 변경으로 App은 토스트를 언제 보여 주는지만 드러내고, 수명주기 규칙은 별도 경계에서 관리한다. 토스트 로직이 다른 화면 오케스트레이션과 섞이지 않아 읽기와 수정 범위가 줄어든다.

추가로 useToastState 테스트를 도입해 표시, 시간 경과 후 exit 전환, exit 완료 시점 처리 경로를 검증했다.

검증:
- npm run test
